feat: differently sized types
totally didn't spend 4 hours debugging a SINGLE missing line.

It reeeeeeaally isn't great, but idc right now.

+++ test/ctest.c
@@ -1,15 +1,15 @@
 #include <stdint.h>
 #include <stdio.h>
 
-extern uint64_t foo(uint64_t a);
-extern uint64_t ayo(uint64_t a, uint64_t b, uint64_t c, uint64_t d, uint64_t e,
-                    uint64_t f, uint64_t g, uint64_t h);
+extern uint32_t foo(uint32_t a);
+extern uint32_t ayo(uint32_t a, uint32_t b, uint32_t c, uint32_t d, uint32_t e,
+                    uint32_t f, uint32_t g, uint32_t h);
 
-uint64_t bar(uint64_t x) { return ayo(x, 1, 2, 3, 4, 5, 6, 7); }
+uint32_t bar(uint32_t x) { return ayo(x, 1, 2, 3, 4, 5, 6, 7); }
 
 int main() {
-  uint64_t f = foo((uint64_t)425); // This will return 420 :)
-  printf("foo(425) = %lu\n", f);
+  uint32_t f = foo((uint32_t)425); // This will return 420 :)
+  printf("foo(425) = %d\n", f);
   return 0;
 }
